refactor: 채팅 보내기 모달 분리
